+++ include/triton/Conversion/TritonGPUToLLVM/TritonGPUToLLVMPass.h
@@ -3,7 +3,6 @@
 
 #include "mlir/Conversion/LLVMCommon/TypeConverter.h"
 #include "mlir/Transforms/DialectConversion.h"
-
 #include <memory>
 
 namespace mlir {
