+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_mlir_conversion_library(TritonGPUToLLVM
-    TritonGPUToLLVM.cpp
+    TritonGPUToLLVMPass.cpp
     PTXAsmFormat.cpp
 
     ADDITIONAL_HEADER_DIRS
