+++ lib/Conversion/TritonGPUToLLVM/PTXAsmFormat.cpp
@@ -1,8 +1,8 @@
+#include "triton/Conversion/TritonGPUToLLVM/PTXAsmFormat.h"
+
 #include "llvm/Support/raw_ostream.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Transforms/DialectConversion.h"
-#include "triton/Conversion/TritonGPUToLLVM/PTXAsmFormat.h"
-
 #include <sstream>
 
 namespace mlir {
